+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This `CHANGELOG.md` implements the spirit of http://keepachangelog.com/.
 
 ### Added
 * Add function base64decode
+* Add function psHeader
 
 ## [1.21](https://github.com/Comcast/eel/compare/v1.20...v1.21) - 2018-07-09